Support **kwargs for requests
Allowing _http_request_request() to receive kwargs allows
the programmer to extend _http_request_requests() so that
all the arguments of python-requests' calls can be supported.
For example, we need to provide the cert argument.
Passing kwargs straight through simplifies the code,
checking is None for every parameter was unecessary.