licapi: ensure version header uses only major version
Using 'version << 8' here resulted in the HTTP request having
a header of:

  version: 255

which is an invalid API version.  This apparently used to work without
error against the license API server but no longer does.  The requests
are now refused with a 404 for any request that includes an invalid API
version.